pnc: add protected attribute support #136

+++ include/as/symbol/function.h
@@ -107,7 +107,7 @@ struct Function : public Scope
 		m_ever_used_like_variable = b;
 	}
 
-	size_t numberOfParameter()
+	unsigned int numberOfParameter()
 	{
 		return m_parameter_types.size();
 	}
